requirements.txt
fastapi
uvicorn
openai
apscheduler
rouge-score
python-dotenv
scikit-learn
numpy
requests
streamlit
plotly
pandas
